requirements.txt
numpy==1.23.4
pandas== 1.5.1
matplotlib == 3.5.3
scikit-learn==1.0.2
joblib == 1.1.1
pillow==7.0.0
seaborn==0.12.0
streamlit
